+++ appli/debimetre.h
@@ -13,10 +13,10 @@
 void DEBIMETRE_init(void);
 void DEBIMETRE_handler(void);
 
-uint16_t DEBIMETRE_get_flow(void);
+uint32_t DEBIMETRE_get_flow(void);
 void DEBIMETRE_set_flow(uint16_t flow);
 
-uint16_t DEBIMETRE_get_stop_value(void);
+uint32_t DEBIMETRE_get_stop_value(void);
 void DEBIMETRE_set_stop_value(uint16_t value);
 
 uint32_t DEBIMETRE_get_consumption(void);
